Update logic for uploading file to s3 storage
- correct file_name for episode
- small refactoring
- remote removing files
- remove tmp file after uploading
